Update patron ineligibility copy logic
