+++ src/components/HoldPages/PatronIneligibilityErrors.tsx
@@ -1,13 +1,18 @@
-import { Text, Box, List } from "@nypl/design-system-react-components"
+import { Box, List, Text } from "@nypl/design-system-react-components"
 
 import type { PatronEligibilityStatus } from "../../types/holdPageTypes"
+import type Item from "../../models/Item"
+
 import RCLink from "../Links/RCLink/RCLink"
 import ExternalLink from "../Links/ExternalLink/ExternalLink"
+import { HoldContactButton } from "./HoldContactButton"
+
 import { PATHS } from "../../config/constants"
 import { appConfig } from "../../config/config"
 
 interface PatronIneligibilityErrorsProps {
   patronEligibilityStatus: PatronEligibilityStatus
+  item: Item
 }
 
 /**
@@ -16,51 +21,61 @@ interface PatronIneligibilityErrorsProps {
  */
 const PatronIneligibilityErrors = ({
   patronEligibilityStatus,
+  item,
 }: PatronIneligibilityErrorsProps) => {
   const { expired, moneyOwed, ptypeDisallowsHolds, reachedHoldLimit } =
     patronEligibilityStatus
 
-  const hasSpecificReason =
-    expired || moneyOwed || ptypeDisallowsHolds || reachedHoldLimit
-
-  // Generic patron error displayed in heading, don't show reasons list if there isn't one
-  if (!hasSpecificReason) return null
-
-  return (
-    <Box mt="xs">
-      <List type="ul" margin={0}>
-        {expired ? (
-          <li>
+  const ineligibilityReasons = [
+    ...(expired
+      ? [
+          <>
             Your account has expired -- Please see{" "}
             <ExternalLink href={appConfig.urls.renewCard}>
               Library Terms and Conditions -- Renewing or Validating Your
               Library Card
             </ExternalLink>{" "}
             about renewing your card.
-          </li>
-        ) : (
-          <></>
-        )}
-        {moneyOwed ? (
-          <li>
+          </>,
+        ]
+      : []),
+    ...(moneyOwed
+      ? [
+          <>
             Your fines have exceeded the limit — you can pay your fines in a
             branch or online from the links under{" "}
             <RCLink href={PATHS.MY_ACCOUNT}>My Account</RCLink>.
-          </li>
-        ) : (
-          <></>
-        )}
-        {ptypeDisallowsHolds ? (
-          <li>Your card does not permit placing holds on ReCAP materials.</li>
-        ) : (
-          <></>
-        )}
-        {reachedHoldLimit ? (
-          <li>You have reached the allowed number of holds.</li>
-        ) : (
-          <></>
-        )}
-      </List>
+          </>,
+        ]
+      : []),
+    ...(ptypeDisallowsHolds
+      ? [<>Your card does not permit placing holds on ReCAP materials.</>]
+      : []),
+    ...(reachedHoldLimit
+      ? [<>You have reached the allowed number of holds.</>]
+      : []),
+  ]
+
+  // Generic patron error displayed in heading, don't show reasons list if there isn't one
+  if (!ineligibilityReasons.length) return null
+
+  return (
+    <Box mt="xs">
+      {ineligibilityReasons.length > 1 ? (
+        <>
+          <List type="ul" margin={0} listItems={ineligibilityReasons} />
+          <Text noSpace mt="xs">
+            Please <HoldContactButton item={item}>contact us</HoldContactButton>{" "}
+            for assistance if required.
+          </Text>
+        </>
+      ) : (
+        <>
+          {ineligibilityReasons.map((reason) => reason)} Please{" "}
+          <HoldContactButton item={item}>contact us</HoldContactButton> for
+          assistance if required.
+        </>
+      )}
     </Box>
   )
 }
